+++ core/lib/refinery/ext/action_view/helpers/form_builder.rb
@@ -1,3 +1,5 @@
+require 'action_view/helpers'
+
 ActionView::Helpers::FormBuilder.module_eval do
 
   def required_label(method, options = {})
